+++ tests/basic.rs
@@ -18,18 +18,18 @@ fn set_size() {
     assert_eq!(parser.screen().size(), (24, 80));
     assert_eq!(parser.screen().cursor_position(), (0, 0));
 
-    parser.set_size(34, 8);
+    parser.screen_mut().set_size(34, 8);
     assert_eq!(parser.screen().size(), (34, 8));
     assert_eq!(parser.screen().cursor_position(), (0, 0));
 
     parser.process(b"\x1b[30;5H");
     assert_eq!(parser.screen().cursor_position(), (29, 4));
 
-    parser.set_size(24, 80);
+    parser.screen_mut().set_size(24, 80);
     assert_eq!(parser.screen().size(), (24, 80));
     assert_eq!(parser.screen().cursor_position(), (23, 4));
 
-    parser.set_size(34, 8);
+    parser.screen_mut().set_size(34, 8);
     assert_eq!(parser.screen().size(), (34, 8));
     assert_eq!(parser.screen().cursor_position(), (23, 4));
 
@@ -37,7 +37,7 @@ fn set_size() {
     assert_eq!(parser.screen().size(), (34, 8));
     assert_eq!(parser.screen().cursor_position(), (0, 0));
 
-    parser.set_size(24, 80);
+    parser.screen_mut().set_size(24, 80);
     assert_eq!(parser.screen().size(), (24, 80));
     assert_eq!(parser.screen().cursor_position(), (0, 0));
 
@@ -45,26 +45,26 @@ fn set_size() {
     assert_eq!(parser.screen().size(), (24, 80));
     assert_eq!(parser.screen().cursor_position(), (23, 4));
 
-    parser.set_size(34, 8);
+    parser.screen_mut().set_size(34, 8);
     parser.process(b"\x1bc01234567890123456789");
     assert_eq!(parser.screen().contents(), "01234567890123456789");
 
-    parser.set_size(24, 80);
+    parser.screen_mut().set_size(24, 80);
     assert_eq!(parser.screen().contents(), "01234567\n89012345\n6789");
 
-    parser.set_size(34, 8);
+    parser.screen_mut().set_size(34, 8);
     assert_eq!(parser.screen().contents(), "01234567\n89012345\n6789");
 
     let mut parser = vt100::Parser::default();
     assert_eq!(parser.screen().size(), (24, 80));
-    parser.set_size(30, 100);
+    parser.screen_mut().set_size(30, 100);
     assert_eq!(parser.screen().size(), (30, 100));
     parser.process(b"\x1b[75Cfoobar");
     assert_eq!(parser.screen().contents(), "                                                                           foobar");
 
     let mut parser = vt100::Parser::default();
     assert_eq!(parser.screen().size(), (24, 80));
-    parser.set_size(30, 100);
+    parser.screen_mut().set_size(30, 100);
     assert_eq!(parser.screen().size(), (30, 100));
     parser.process(b"1\r\n2\r\n3\r\n4\r\n5\r\n6\r\n7\r\n8\r\n9\r\n10\r\n11\r\n12\r\n13\r\n14\r\n15\r\n16\r\n17\r\n18\r\n19\r\n20\r\n21\r\n22\r\n23\r\n24\x1b[24;99Hfoobar");
     assert_eq!(parser.screen().contents(), "1\n2\n3\n4\n5\n6\n7\n8\n9\n10\n11\n12\n13\n14\n15\n16\n17\n18\n19\n20\n21\n22\n23\n24                                                                                                foobar");